fixes responiveness issues on smaller devices and code refactors
