+++ src/App.tsx
@@ -1,5 +1,6 @@
 import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
 import { ThemeProvider } from 'styled-components';
+import { AppLayout } from './components/app-layout';
 import {
   ClaimDetailsScreen,
   ClaimsScreen,
@@ -17,13 +18,15 @@ function App() {
     <ThemeProvider theme={theme}>
       <Router>
         <Switch>
-          <Route path='/' exact component={HomeScreen} />
-          <Route path='/login' exact component={LoginScreen} />
-          <Route path='/register' exact component={RegisterScreen} />
-          <Route path='/createsafe' exact component={CreateSafeScreen} />
-          <Route path='/claims' exact component={ClaimsScreen} />
-          <Route path='/claims/:id' exact component={ClaimDetailsScreen} />
-          <Route path='/profile' exact component={ProfileScreen} />
+          <AppLayout>
+            <Route path='/' exact component={HomeScreen} />
+            <Route path='/login' exact component={LoginScreen} />
+            <Route path='/register' exact component={RegisterScreen} />
+            <Route path='/createsafe' exact component={CreateSafeScreen} />
+            <Route path='/claims' exact component={ClaimsScreen} />
+            <Route path='/claims/:id' exact component={ClaimDetailsScreen} />
+            <Route path='/profile' exact component={ProfileScreen} />
+          </AppLayout>
         </Switch>
         <GlobalStyle />
       </Router>
