+++ src/components/app-layout/app-layout.component.props.ts
@@ -1,0 +1,6 @@
+/**
+ *  Interface for AppLayout which accepts Children as props.
+ */
+export interface AppLayoutProps extends React.HTMLAttributes<HTMLDivElement> {
+  children?: React.ReactNode;
+}
