+++ src/components/primitive/navbar/navbar.component.props.ts
@@ -1,15 +1,5 @@
-import { IconProps } from './../icon-svg/icon-svg.component.props';
-import { TextComponentProps } from './../text/text.component.props';
-
-/**
- * Different screen sizes used across the screens.
- */
-export enum BreakPoints {
-  small = '480px',
-  medium = '768px',
-  large = '1040px',
-  xLarge = '1200px',
-}
+import { IconProps } from '../icon-svg/icon-svg.component.props';
+import { TextComponentProps } from '../text/text.component.props';
 
 /**
  * Interface for Navbar items.
