+++ src/components/primitive/navbar/navbar.component.styles.ts
@@ -1,31 +1,22 @@
 import styled from 'styled-components';
-import { Navlink, Box } from '../../primitive';
-import { BreakPoints } from './navbar.component.props';
+import { Box, Navlink } from '../../primitive';
+import { BreakPoints } from '../../../utils';
 
 /**
- * Styling for sidebar.
+ * Styling for SideBarWrapper.
  */
-export const SideBar = styled(Box)`
+export const SideBarWrapper = styled.div`
   display: flex;
   flex-direction: column;
   flex: 1;
   background-color: ${({ theme: { colors } }) => colors.white};
-  padding: 2rem;
-  max-width: 34.3rem !important;
-  height: 100vh;
 
-  /**
- * Styles that are applied on medium screens.
- */
   @media screen and (max-width: ${BreakPoints.medium}) {
-    display: none !important;
-  }
-
-  /**
- * Styles that are applied on medium screens and small screens.
- */
-  @media screen and (max-width: ${BreakPoints.medium}) and (max-width: ${BreakPoints.small}) {
-    display: none !important;
+    flex-direction: row;
+    height: 7rem;
+    width: 100vw;
+    position: fixed !important;
+    bottom: 0;
   }
 `;
 
@@ -37,23 +28,27 @@ export const LogoWrapper = styled.div`
   padding-left: 3rem;
   margin: 3rem 0 4rem 0 !important;
 
-  /**
- * Styles that are applied on medium screens.
- */
   @media screen and (max-width: ${BreakPoints.medium}) {
     display: none !important;
   }
 
-  /**
- * Styles that are applied on medium screens and small screens.
- */
   @media screen and (max-width: ${BreakPoints.medium}) and (max-width: ${BreakPoints.small}) {
     display: none !important;
   }
 `;
 
+/**
+ * Styling for NavLabelContainer.
+ */
 export const NavLabelContainer = styled(Box)`
   padding: 1rem !important;
+
+  @media screen and (max-width: ${BreakPoints.medium}) {
+    padding: 0 !important;
+    display: flex;
+    flex-direction: row !important;
+    justify-content: space-evenly;
+  }
 `;
 
 /**
