+++ src/components/primitive/navbar/navbar.component.tsx
@@ -1,14 +1,14 @@
-import { useHistory } from 'react-router';
-import { navItems } from './navbar';
+import { useHistory } from 'react-router-dom';
 import { Image } from '../../primitive';
 import { safientLogo } from '../../../assets';
-import { SideBar, LogoWrapper, NavLabelContainer, NavLabel } from './navbar.component.styles';
+import { navItems } from './navbar';
+import { SideBarWrapper, LogoWrapper, NavLabelContainer, NavLabel } from './navbar.component.styles';
 
 export const NavBar: React.FC = () => {
   const history = useHistory();
 
   return (
-    <SideBar>
+    <SideBarWrapper>
       <LogoWrapper>
         <Image src={safientLogo} width={12.1} />
       </LogoWrapper>
@@ -17,6 +17,6 @@ export const NavBar: React.FC = () => {
           return <NavLabel key={index} active={item.path === history.location.pathname} {...item} />;
         })}
       </NavLabelContainer>
-    </SideBar>
+    </SideBarWrapper>
   );
 };
