+++ src/utils/spacing.ts
@@ -3,18 +3,10 @@
  */
 
 export const spacing = {
-  tiny: 1,
-  small: 2,
-  medium: 3,
-  large: 4,
-  xTiny: 0.8,
-  xxTiny: 0.4,
-  xSmall: 1.8,
-  xxSmall: 1.6,
-  xMedium: 2.6,
-  xxMedium: 2.2,
-  xLarge: 3.6,
-  xxLarge: 3.2,
+  tiny: 0.6,
+  small: 1.9,
+  medium: 2.2,
+  large: 2.6,
 };
 
 export type SpacingTypes = keyof typeof spacing;
