+++ src/utils/variables.ts
@@ -1,20 +1,9 @@
-export const SHADOWS = {
-  default: '0px 16px 32px rgba(221, 230, 237, 0.4)',
-  small: '0 5px 10px rgba(0,0,0,0.12)',
-  medium: '0 8px 30px rgba(0,0,0,0.12)',
-  large: '0 30px 60px rgba(0,0,0,0.12)',
-};
-
-export const SCREEN_SIZE = {
-  SM: '576px', // phones in min-width
-  MD: '768px', // tablets
-  LG: '1040px', // laptops/desktops
-  XL: '1200px', // extra Large laptops/desktops
-};
-
-export const MEDIA_QUERIES = {
-  sm: 'only screen and (max-width:576px)',
-  md: 'only screen and (max-width:768px)',
-  lg: 'only screen and (max-width:1040px)',
-  xl: 'only screen and (max-width:1200px)',
-};
+/**
+ * Different screen sizes used across the screens.
+ */
+export enum BreakPoints {
+  small = '480px',
+  medium = '768px',
+  large = '1040px',
+  xLarge = '1200px',
+}
